Make roomrateperday on new installations longer so that countries like Columbia who have prices in the millions of pesos can still list prices.

+++ database/tables/jomres_rates.php
@@ -27,7 +27,7 @@ CREATE TABLE IF NOT EXISTS `#__jomres_rates` (
 	`rate_description` TEXT,
 	`validfrom` CHAR(10),
 	`validto` CHAR(10),
-	`roomrateperday` DOUBLE(11,7) NOT NULL DEFAULT 0,
+	`roomrateperday` DOUBLE(20,7) NOT NULL DEFAULT 0,
 	`extra_guests_price` DOUBLE(11,2) NOT NULL DEFAULT 0,
 	`modifiers` TEXT,
 	`mindays` SMALLINT(5),
